pyproject.toml
[project]
name = "python-graphql-example"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10.7"
dependencies = [
    "fastapi>=0.115.6",
    "psycopg2-binary>=2.9.10",
    "requests>=2.32.3",
    "sqlalchemy>=2.0.36",
    "strawberry-graphql[fastapi]>=0.253.1",
    "uvicorn>=0.32.1",
]

[dependency-groups]
dev = [
    "black>=24.10.0",
    "isort>=5.13.2",
    "pytest>=8.3.4",
]
